+++ spec/lib/glimmer/libui/custom_control/refined_table_spec.rb
@@ -74,6 +74,12 @@ RSpec.describe Glimmer::LibUI::CustomControl::RefinedTable do
       expect(result).to be_falsey
     end
     
+    it 'does not match a specific column by middle letter of column name' do
+      query = 'n:john' # attempting to match first name with n
+      result = described_class::FILTER_DEFAULT.call(row_hash, query)
+      expect(result).to be_falsey
+    end
+    
     it 'matches first specific column in multiple having first letter of column name' do
       query = 'c:champaign' # city
       result = described_class::FILTER_DEFAULT.call(row_hash, query)
@@ -110,7 +116,17 @@ RSpec.describe Glimmer::LibUI::CustomControl::RefinedTable do
       expect(result).to be_falsey
     end
     
-    # TODO test double quoted column and value for specific column
+    it 'matches a specific column by full double-quoted column name and double-quoted column value' do
+      query = '"first name":"jo n"'
+      result = described_class::FILTER_DEFAULT.call(row_hash, query)
+      expect(result).to be_truthy
+    end
+    
+    it 'does not match a specific column by full double-quoted column name and double-quoted column value' do
+      query = '"first name":"jo e"'
+      result = described_class::FILTER_DEFAULT.call(row_hash, query)
+      expect(result).to be_falsey
+    end
     
     it 'matches a double-quoted column-specific term, multiple words, a double-quoted term' do
       query = ' "First Name":Jo Doe Illinois "a-Champ" '
@@ -140,6 +156,12 @@ RSpec.describe Glimmer::LibUI::CustomControl::RefinedTable do
       query = 'First_name:Jos Doe Illinois "a-Champ"'
       result = described_class::FILTER_DEFAULT.call(row_hash, query)
       expect(result).to be_falsey
+    end
+    
+    it 'matches column-specifc term without value, treating value as empty string' do
+      query = 'first_name:'
+      result = described_class::FILTER_DEFAULT.call(row_hash, query)
+      expect(result).to be_truthy
     end
   end
 end